Agregando comprobante y cambios en interfaz (Swal)

+++ src/layouts/productos/tipo_productos/index.js
@@ -17,6 +17,7 @@ import Icon from "@mui/material/Icon";
 // Data
 //import authorsTableData from "layouts/productos/tipo_productos/data/authorsTableData";
 import { useEffect, useState } from "react";
+import Swal from "sweetalert2";
 
 function Tables() {
   //const { columns, rows } = authorsTableData();
@@ -43,6 +44,7 @@ function Tables() {
       );
 
       const data = await response.json();
+      console.log(data);
       setProducts(data);
     } catch (error) {
       console.error("Error de red:", error);
@@ -101,8 +103,17 @@ function Tables() {
         );
 
         if (response.ok) {
-          document.getElementById("btnCerrar").click();
-          getProducts();
+          Swal.fire({
+            title: "Edicion Exitosa!",
+            text: "Tipo de producto editado exitosamente",
+            icon: "success",
+            timer: 500,
+            position: "top-end",
+            toast: true,
+          }).then((result) => {
+            document.getElementById("btnCerrar").click();
+            getProducts();
+          });
         } else {
           console.log("Error al editar");
         }
@@ -125,8 +136,17 @@ function Tables() {
         );
 
         if (response.ok) {
-          document.getElementById("btnCerrar").click();
-          getProducts();
+          Swal.fire({
+            title: "Edicion Exitosa!",
+            text: "Tipo de producto agregado exitosamente",
+            icon: "success",
+            timer: 500,
+            position: "top-end",
+            toast: true,
+          }).then((result) => {
+            document.getElementById("btnCerrar").click();
+            getProducts();
+          });
         } else {
           console.log("Error al agregar");
         }
@@ -158,8 +178,20 @@ function Tables() {
   };
 
   const deleteProduct = async (id) => {
-    onDeleteProduct(id);
-    getProducts();
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        onDeleteProduct(id);
+        getProducts();
+      }
+    });
   };
 
   return (
@@ -182,6 +214,7 @@ function Tables() {
                 <MDTypography variant="h6" color="white">
                   Tipo de productos
                   <button
+                    style={{ marginLeft: "20px" }}
                     onClick={() => openModal(1)}
                     className="btn btn-dark"
                     data-bs-toggle="modal"
@@ -251,7 +284,7 @@ function Tables() {
           </Grid>
         </Grid>
       </MDBox>
-      <div id="modalProducts" className="modal fade" aria-hidden="true" style={{ zIndex: "9999" }}>
+      <div id="modalProducts" className="modal fade" aria-hidden="true">
         <div className="modal-dialog modal-dialog-centered">
           <div className="modal-content">
             <div className="modal-header">
